backend/requirements.txt
fastapi
redis
python-dotenv
uvicorn
